fix: unblock wasm32 workspace test builds
Fixes a series of wasm32-unknown-unknown compilation failures when running
`cargo test --target wasm32-unknown-unknown --workspace --tests --no-run --locked`.

- aero-gpu: skip wgpu-backed unit tests on wasm32; validate persistent shader reflection via sampler_dim_key
- aero-d3d11: handle new IR variants (Itof/Ftoi/Ftou) and ISubC dst_carry rename
- aero-d3d9: handle Dp2Add in SM3 pixel input scan
- aero-storage: skip qcow2/vhd format integration test on wasm32 (VirtualDisk is !Send there)
- aero-platform: restore lapics_iter helper used by interrupt router
- emulator: avoid calling VgaDevice::active_scanout_update without wasm atomics
- aero-usb: avoid shifting usize by 32 in xHCI tests on wasm32
- device-contract-validator/aero-d3d9: fix test compilation issues (new validate arg, duplicate helper)
- conformance: cfg-gate host-only signal-based tests on wasm32
- aero-d3d11 runtime: keep shader cache enums in sync with new stages/bindings

Result: wasm32 workspace test build now compiles.

+++ crates/aero-usb/tests/xhci_ports.rs
@@ -215,7 +215,7 @@ fn xhci_tick_advances_mfindex_and_flushes_events_after_erst_configured() {
         &mut mem,
         regs::REG_INTR0_ERDP_HI,
         4,
-        (ring_base >> 32) as u32,
+        ((ring_base as u64) >> 32) as u32,
     );
     xhci.mmio_write(&mut mem, regs::REG_INTR0_IMAN, 4, IMAN_IE);
